+++ frontend/src/layouts/MainLayout.vue
@@ -25,10 +25,8 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, ref } from 'vue';
+import { defineComponent, provide, ref } from 'vue';
 import EssentialLink from '../components/EssentialLink.vue';
-import gql from 'graphql-tag';
-import { useMutation } from 'villus';
 
 const linksList = [
   {
@@ -47,19 +45,6 @@ const linksList = [
   },
 ];
 
-const TOKEN_MUTATION = gql`
-  mutation CheckToken($token: String!) {
-    checkToken(token: $token) {
-      username
-      familyName
-      givenName
-      gender
-      birthDay
-      token
-    }
-  }
-`;
-
 export default defineComponent({
   name: 'MainLayout',
 
@@ -67,23 +52,16 @@ export default defineComponent({
     EssentialLink,
   },
 
-  async setup() {
+  props: {
+    userMutation: {
+      type: Function,
+      required: true,
+    },
+  },
+
+  async setup(props) {
     const leftDrawerOpen = ref(false);
-    const { execute } = useMutation(TOKEN_MUTATION, {});
-    const variables = {
-      token: localStorage.getItem('token'),
-    };
-
-    const fetchData = async () => {
-      try {
-        const response = await execute(variables);
-        return response.data.checkToken;
-      } catch (error) {
-        return '';
-      }
-    };
-
-    const user = await fetchData();
+    const user = await props.userMutation();
 
     return {
       essentialLinks: linksList,
