--- frontend/src/script/useAuth.ts
@@ -1,11 +1,0 @@
-import { computed } from 'vue';
-
-export default function useAuth() {
-  const isLoggedIn = computed(() => {
-    return !!localStorage.getItem('token');
-  });
-
-  return {
-    isLoggedIn,
-  };
-}
